refactor : convert 로직 util 분리
